requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
Deprecated==1.2.13
Django==3.2.10
django-debug-toolbar==3.2.4
django-filter==21.1
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
wrapt==1.13.3
